Report for identifications

Generates a report for identifications

Usage

get_ID_Report(input_list)

Arguments

input_list

A list with data frames and respective level information.

Value

This function returns a list. For each analysis a respective data frame including number of identifications per run is stored in the generated list.

Analysis - analysis name.

Run - run information.

Precursor.IDs - number of precursor identification.

Peptide.IDs - number of peptide identification.

Protein.IDs - number of protein identification.

ProteinGroup.IDs - number of proteingroup identification.

Details

For each submitted data a report with achieved number of identifications is generated on precursor-, peptide-, protein- and proteingroup-level.

Author

Oliver Kardell

Examples

# Load libraries
library(tibble)
library(stringr)

# Example data
data <- list(
DIANN = list(
  filename = "B",
  software = "DIA-NN",
  data = list(
     "DIA-NN" = tibble::tibble(
        Run_mpwR = rep(c("A","B"), times = 10),
        Precursor.IDs_mpwR = rep(c("A2", "A3", "B2", "B3", "C1"), each = 4),
        Protein.IDs_mpwR = rep(c("A2", "A3", "B2", "B3", "C1"), each = 4),
        Peptide.IDs_mpwR = rep(c("A", "A", "B", "B", "C"), each = 4),
        ProteinGroup.IDs_mpwR = rep(c("A2", "A3", "B2", "B3", "C1"), each = 4)
     )
  )
)
)

# Result
output <- get_ID_Report(
  input_list = data
)