Load and Prepare the input data

Input data will be imported, renamed and default filtering will be applied

Usage

prepare_mpwR(path, diann_addon_pg_qval = 0.01, diann_addon_prec_qval = 0.01)

Arguments

path

Path to folder where the input data is stored - only input data. No subfolders or other files. Analysis name as prefix + for MaxQuant: _evidence, _peptides, _proteinGroups; for PD - R-friendly headers enabled: _PSMs, _Proteins, _PeptideGroups, _ProteinGroups; for DIA-NN, Spectronaut and Generic: _Report

diann_addon_pg_qval

Numeric between 0 and 1. Applied only to DIA-NN data: diann_addon_pg_qval <= PG.Q.Value.

diann_addon_prec_qval

Numeric between 0 and 1. Applied only to DIA-NN data: diann_addon_prec_qval <= Q.Value.

Value

A list - each list entry has filename and software info as well as stored data.

Details

Function for easily importing the default software outputs and preparing for downstream analysis with mpwR within one folder. As default for MaxQuant "Reverse", "Potential contaminants" and "Only identified by site" are filtered out. As default for PD only "High" confidence identifications are included and for Found in Sample column(s) also only "High" identifications. Contaminants are filtered out. As default for Spectronaut only EG.Identified equals TRUE are included.

Author

Oliver Kardell

Examples

if (FALSE) { # \dontrun{
prepare_mpwR(path = "DIRECTORY_TO_FILES")
} # }