Create generic template

Generation of a template.csv file for generic input data. The template is provided in long-format.

Usage

write_generic_template(path_filename)

Arguments

path_filename

Path to folder where template is generated and user-defined filename

Value

This function returns a csv-file with the following columns:

Run_mpwR - name of file(s).

ProteinGroup.IDs_mpwR - ProteinGroup with identifier(s) of protein(s) contained in the protein group.

Protein.IDs_mpwR - Protein identifier(s).

Peptide.IDs_mpwR - Sequence representation plus possible post-translational modifications.

Precursor.IDs_mpwR - Sequence representation plus possible post-translational modifications including charge state.

Stripped.Sequence_mpwR - The amino acid sequence of the identified peptide without modifications.

Precursor.Charge_mpwR - Charge state of the precursor.

Missed.Cleavage_mpwR - Number of missed enzymatic cleavages.

Retention.time_mpwR - Retention time in minutes in the elution profile of the precursor ion.

ProteinGroup_LFQ_mpwR - LFQ intensity column on proteingroup-level

Peptide_LFQ_mpwR - LFQ intensity column on petide-level

Details

The generated template.csv file can be used to create a software-independent input file for mpwR. Example entries are provided. The template file - filename_Report.csv - is generated. The appendix "_Report" is required for importing with mpwR. Note that the template is in long-format, so each ProteinGroup.ID has possible multiple entries depending on the number of Precursor.IDs.

Author

Oliver Kardell

Examples

if (FALSE) { # \dontrun{
write_generic_template(path = "DIRECTORY_WHERE_FILE_IS_GENERATED/filename")
} # }